% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{chol.abc}
\alias{chol.abc}
\title{2D Cholesky decomposition}
\usage{
\method{chol}{abc}(v1, v2, v12)
}
\arguments{
\item{v1}{element of Sigma}

\item{v2}{element of Sigma}

\item{v12}{element of Sigma}
}
\value{
elements a, b, c of Cholesky decomposition
}
\description{
Closed form Cholesky of 2x2 variance matrices
}
\details{
Let Sigma=((v1 v12), (v12 v2))
Calculate A st A=((a 0),(b c)) and AA'=S
}
\author{
Chris Wallace
}
