% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rbivnorm}
\alias{rbivnorm}
\alias{dbivnorm}
\title{bivariate random variables with non-constant mean and var}
\usage{
rbivnorm(n, mean = rep(0, 2), v1 = 1, v2 = 1, v12 = 0, ...)

dbivnorm(x, mean = rep(0, 2), v1 = 1, v2 = 1, v12 = 0, ...)
}
\arguments{
\item{n}{number of replicates}

\item{mean}{mean of the ditribution}

\item{v1}{element of Sigma}

\item{v2}{element of Sigma}

\item{v12}{element of Sigma}

\item{...}{passed to dmvnorm}

\item{x}{matrix of quantiles}
}
\value{
matrix of random samples


}
\description{
rmvnorm in bivariate case
}
\details{
Let Sigma=((v1 v12)', (v12 v2)'), mu=(m1 m2)'

If Z ~ N(0,I2) is the bivariate standard normal, then
Y = A \%*\% Z + mu is a bivariate normal with mean mu and variance
Sigma = AA'
}
\author{
Chris Wallace

Chris Wallace
}
